Refactor some code into 'fft_conv_pytorch/utils.py' as a shared resource
